+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/AnimatedArrow/AnimatedArrow.tsx
@@ -1,5 +1,5 @@
-import { MapPageConfig } from '../../../../../constants';
-import materialIcon from '../../../../../utils/materialIcon';
+import { MapPageConfig } from '../../../../../../../constants';
+import materialIcon from '../../../../../../../utils/materialIcon';
 
 export const AnimatedArrow: React.FC = () => {
   return (
