+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/Marker/index.ts
