+++ apps/web/src/app/(needLogin)/main/_components/MapPage/index.ts
@@ -1,0 +1,1 @@
+export * from './MapPage';
