+++ apps/web/src/app/(needLogin)/main/_components/index.ts
@@ -1,0 +1,1 @@
+export * from './MapPage';
