+++ apps/web/src/app/(needLogin)/main/page.tsx
@@ -1,0 +1,11 @@
+import { MapPage } from './_components';
+
+const MainPage: React.FC = () => {
+  return (
+    <div>
+      <MapPage />
+    </div>
+  );
+};
+
+export default MainPage;
